Safari-611.1.10/LayoutTests/fast/selectors/nth-last-child-with-backtracking-expected.txt

Test nth-last-child inside deep backtracking.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll("li+li+:nth-last-child(-5n+14)+li+li+li~li+li+li~li").length is 7
PASS allItems.length is 20
PASS coloredItems is 7
PASS document.querySelectorAll("ul>[foo=bar]:nth-last-child(-3n+18)+li+li+li~li+li+li~li>span.first>span.second a").length is 11
PASS allLinks.length is 20
PASS backgroundColoredCount is 11
PASS document.querySelectorAll("ul>:nth-last-child(-3n+3)>*>* a").length is 1
PASS nonOpaqueCount is 1
PASS successfullyParsed is true

TEST COMPLETE
